Install ThemisPP --with-cpp
Add an option to install ThemisPP support when installing Themis.
Options are frowned upon in the main Homebrew repository [1], but
we're not there. It would be a waste to have two separate packages
that download the same source tarball. In case of ThemisPP we don't
even need to compile anything, that's just copying C++ headers (and
you need to install Themis Core for them anyway).

[1]: https://github.com/Homebrew/homebrew-core/issues/31510