Install Themis JNI --with-java
Another option to install Themis JNI library which requires Themis Core
to operate. Now we do need to compile something.

Java is not available as Homebrew package in the main repo. However,
it's possible to install it via Casks. Unfortunately, we cannot depend
on casks, so it's assumed that the system has Java installed when using
the --with-java option.

Currently Themis JNI library is tightly coupled with the core library so
they have to compiled together using the same private header files.
Therefore we use options, not a separate package.